requirements.txt
onnxruntime
numpy
pillow